pyproject.toml
[project]
name = "manga-panel-split"
version = "0.1.0"
description = ""
authors = [
    {name = "a1528zhang",email = "beingszhang@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "transformers (==4.38.0)",
    "torch (>=2.10.0,<3.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "numpy (>=2.4.2,<3.0.0)"
]

[tool.poetry]
package-mode = false
packages = [{include = "manga_panel_split", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
